test: consolidate extension-registry reset into conftest.py
_reset_registry cleared varying subsets of the four
app.platform.extensions globals (_extensions, _routers, _loaded,
_slot_owners) across 13 files: 7 cleared only two of the four, 3 added
_slot_owners, and 3 worker-bootstrap files cleared all four. A partial
reset is a latent cross-test pollution bug waiting for a file to start
touching a slot its own copy-pasted subset didn't anticipate.

Move one _reset_registry clearing all four globals into conftest.py,
plus the entry_points-patching _clean_registry fixture that 8 of the
13 files shared verbatim (module-level pytestmark replaces each
file's local autouse fixture). The two files whose _clean_registry
does not patch entry_points, and the three worker-bootstrap files
with their own combined registry+edition reset fixtures, keep their
local fixtures but now import the shared _reset_registry.
